refactoring etl folder structure
