+++ src/tests/core/engine/processors/abstract_etl_processor_test.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta
 from unittest import mock
 
-from src.core.engine.processors.abstract_etl_processor import AbstractEtl
+from src.core.engine.processors.etl.abstract_etl_processor import AbstractEtl
 from src.tests.base_env_config_test import BaseEnvConfigTest
 from src.tests.utils.fixtures.app_processor_config_fixture import (
     build_app_processor_config,
