+++ src/tests/core/engine/processors/mongo_to_postgres_order_status_history_test.py
@@ -1,9 +1,9 @@
 from unittest import mock
 
 from src.constants.audit import InternalUsers
-from src.core.engine.processors.abstract_etl_processor import AbstractEtl
-from src.core.engine.processors.mongo_to_postgresql_order_status_history import (
-    MongoToPostgresqlOrderStatusHistory,
+from src.core.engine.processors.etl.abstract_etl_processor import AbstractEtl
+from src.core.engine.processors.etl.mongo_to_postgresql_order_status_history_etl import (
+    MongoToPostgresqlOrderStatusHistoryEtl,
 )
 from src.tests.lib.repositories.mongo_engine_base_repository_impl_test import (
     MongoEngineBaseRepositoryTestCase,
@@ -21,7 +21,7 @@ class MongoToPostgresOrderStatusHistoryTest(MongoEngineBaseRepositoryTestCase):
     def after_base_setup(self):
         self.app_config = build_app_processor_config()
 
-        self.mongo_to_postgres_etl = MongoToPostgresqlOrderStatusHistory(AbstractEtl)
+        self.mongo_to_postgres_etl = MongoToPostgresqlOrderStatusHistoryEtl(AbstractEtl)
 
         etl = self.mongo_to_postgres_etl
         etl.app_processor_config = self.app_config
